+++ include/ui/setup.hpp
@@ -290,9 +290,16 @@ namespace nwo5::ui {
             return {m_node};
         }
 
-        template<typename T>
-        Setup id(T pID) requires std::derived_from<Node, cocos2d::CCNode> {
+        Setup id(const std::string& pID) requires std::derived_from<Node, cocos2d::CCNode> {
             m_node->setID(pID);
+            return {m_node};
+        }
+        Setup id(geode::ZStringView pID) requires std::derived_from<Node, cocos2d::CCNode> {
+            m_node->setID(pID);
+            return {m_node};
+        }
+        Setup id(std::string_view pID) requires std::derived_from<Node, cocos2d::CCNode> {
+            m_node->setID(std::string{pID});
             return {m_node};
         }
         template <typename T, typename... Args>